Tensorflow 1.0 code is now Python 3 compatible

+++ code1.0/dgp_rff.py
@@ -19,12 +19,10 @@ import numpy as np
 import matplotlib
 matplotlib.use('Agg')
 import matplotlib.pyplot as plt
-from dataset import DataSet
-import utils
-import likelihoods
 import time
-import dgp_interface
-
+from .dataset import DataSet
+from . import utils
+from . import dgp_interface
 
 current_milli_time = lambda: int(round(time.time() * 1000))
 
@@ -81,7 +79,7 @@ class DgpRff(dgp_interface.DGPRFF_Interface):
 
             ## Apply the activation function corresponding to the chosen kernel - PHI
             if self.kernel_type == "RBF":
-                Phi = tf.exp(0.5 * self.log_theta_sigma2[i]) / (tf.sqrt(1. * self.n_rff[i])) * tf.concat(axis=2, values=[tf.cos(layer_times_Omega), tf.sin(layer_times_Omega)])
+                Phi = tf.exp(0.5 * self.log_theta_sigma2[i]) / (tf.sqrt(tf.cast(1. * self.n_rff[i], tf.float32))) * tf.concat(axis=2, values=[tf.cos(layer_times_Omega), tf.sin(layer_times_Omega)])
             if self.kernel_type == "arccosine":
                 if self.arccosine_degree == 0:
                     Phi = tf.exp(0.5 * self.log_theta_sigma2[i]) / (tf.sqrt(1. * self.n_rff[i])) * tf.concat(axis=2, values=[tf.sign(tf.maximum(layer_times_Omega, 0.0))])
